fix(ui): normalize Lighthouse MCP tool allowlist
- Move Lighthouse MCP tool allowlist into a shared module

- Normalize legacy prowler_app tool names before allowlist checks

- Add unit coverage for current and legacy tool namespaces
